src/portal_automation/requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
pytest>=7.0.0
python-dotenv>=0.19.0
faker>=18.0.0
pytest-html>=3.1.0
